+++ src/posix/platform/radio_spinel.hpp
@@ -37,12 +37,13 @@
 #include <openthread/platform/radio.h>
 
 #include "frame_queue.hpp"
-#include "hdlc.hpp"
+#include "hdlc_interface.hpp"
 #include "spinel.h"
 
 namespace ot {
+namespace PosixApp {
 
-class RadioSpinel
+class RadioSpinel : public HdlcInterface::Callbacks
 {
 public:
     /**
@@ -439,10 +440,19 @@ public:
     otError PlatDiagProcess(const char *aString, char *aOutput, size_t aOutputMaxLen);
 #endif
 
+    /**
+     *  This method processes a received Spinel frame.
+     *
+     * @param[in] aBuffer  A pointer to buffer containing the frame.
+     * @param[in] aLength  Length (number of bytes) in the received frame.
+     *
+     */
+    void HandleSpinelFrame(const uint8_t *aFrame, uint16_t aLength);
+
 private:
     enum
     {
-        kMaxSpinelFrame    = 2048, ///< Max size in bytes for transferring spinel frames.
+        kMaxSpinelFrame    = HdlcInterface::kMaxFrameSize,
         kMaxWaitTime       = 2000, ///< Max time to wait for response in milliseconds.
         kVersionStringSize = 128,  ///< Max size of version string.
         kCapsBufferSize    = 100,  ///< Max buffer size used to store `SPINEL_PROP_CAPS` value.
@@ -451,9 +461,6 @@ private:
     otError CheckSpinelVersion(void);
     otError CheckCapabilities(void);
     otError CheckRadioCapabilities(void);
-    void    DecodeHdlc(const uint8_t *aData, uint16_t aLength);
-    void    ReadAll(void);
-    otError WriteAll(const uint8_t *aBuffer, uint16_t aLength);
     void    ProcessFrameQueue(void);
 
     /**
@@ -531,7 +538,6 @@ private:
     {
         static_cast<RadioSpinel *>(aContext)->HandleSpinelFrame(aBuffer, aLength);
     }
-    void HandleSpinelFrame(const uint8_t *aBuffer, uint16_t aLength);
 
     /**
      * This method returns if the property changed event is safe to be handled now.
@@ -546,7 +552,7 @@ private:
      */
     bool IsSafeToHandleNow(spinel_prop_key_t aKey) const
     {
-        return !((mIsDecoding || mWaitingKey != SPINEL_PROP_LAST_STATUS) &&
+        return !((mHdlcInterface.IsDecoding() || mWaitingKey != SPINEL_PROP_LAST_STATUS) &&
                  (aKey == SPINEL_PROP_STREAM_RAW || aKey == SPINEL_PROP_MAC_ENERGY_SCAN_RESULT));
     }
 
@@ -562,6 +568,8 @@ private:
 
     otInstance *mInstance;
 
+    HdlcInterface mHdlcInterface;
+
     uint16_t          mCmdTidsInUse;    ///< Used transaction ids.
     spinel_tid_t      mCmdNextTid;      ///< Next available transaction id.
     spinel_tid_t      mTxRadioTid;      ///< The transaction id used to send a radio frame.
@@ -572,10 +580,7 @@ private:
     uint32_t          mExpectedCommand; ///< Expected response command of current transaction.
     otError           mError;           ///< The result of current transaction.
 
-    uint8_t       mHdlcBuffer[kMaxSpinelFrame];
-    Hdlc::Decoder mHdlcDecoder;
-    Hdlc::Encoder mHdlcEncoder;
-    FrameQueue    mFrameQueue;
+    FrameQueue mFrameQueue;
 
     uint8_t       mRxPsdu[OT_RADIO_FRAME_MAX_SIZE];
     uint8_t       mTxPsdu[OT_RADIO_FRAME_MAX_SIZE];
@@ -593,10 +598,8 @@ private:
     otError      mTxError;
     char         mVersion[kVersionStringSize];
 
-    int          mSockFd;
     otRadioState mState;
     bool         mIsAckRequested : 1;    ///< Ack requested.
-    bool         mIsDecoding : 1;        ///< Decoding hdlc frames.
     bool         mIsPromiscuous : 1;     ///< Promiscuous mode.
     bool         mIsReady : 1;           ///< NCP ready.
     bool         mSupportsLogStream : 1; ///< RCP supports `LOG_STREAM` property with OpenThread log meta-data format.
@@ -608,6 +611,7 @@ private:
 #endif
 };
 
+} // namespace PosixApp
 } // namespace ot
 
 #endif // RADIO_SPINEL_HPP_
